refactor bench runner into separate files
